requirements.txt
alembic==1.16.4
asyncpg==0.30.0
bcrypt==4.0.1
fastapi==0.116.1
gunicorn==23.0.0
loguru==0.7.3
passlib[bcrypt]
pre-commit==4.3.0
pydantic[email]
pydantic-settings==2.10.1
PyJWT==2.10.1
redis==6.4.0
SQLAlchemy==2.0.43
uvicorn==0.35.0
